chore: project color as constant
